\name{datatypes}
\alias{datatypes}
\docType{data}
\title{
Supported Datatypes
}
\description{
This dataframe contains details of the supported datatypes in FIACH.
}
\usage{data(datatypes)}
\format{
  A data frame with 14 observations on the following 3 variables.
  \describe{
    \item{\code{niftyCodes}}{Strings used to define RNiftyReg datatype output}
    \item{\code{niftiCodes}}{Codes used to define nifti datatypes}
    \item{\code{bitpix}}{number of bits per pixel}
    
  }
}

\examples{
data(datatypes)
datatypes
}
\keyword{datasets}
