\name{fiachSplit}
\alias{fiachSplit}
\title{
Split 4D .nii files into 3D .nii files
}
\description{
This function splits 4D .nii files into 3D .nii files.}
\usage{
fiachSplit(input)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{input}{
Character string specifying filepath to 4D file.}
}
\value{
Nothing is returned to R but the 3D files are written to the directory from which the 4D file came from.
}

\examples{
## to do
}
